--- src/shared/entities/type.ts
@@ -1,9 +1,0 @@
-export const CATEGORY = {
-  HOT_PLACE: '핫플',
-  FAMOUS_RESTAURANT: '맛집',
-  ACTIVITY: '액티비티',
-  CULTURE_AND_ART: '문화/예술',
-  ETC: '기타',
-}
-
-export type CategoryType = keyof typeof CATEGORY
