+++ src/widgets/course-plan-layout/index.tsx
@@ -3,7 +3,7 @@ import PlaceCollapse from '@/src/shared/ui/PlaceCollapse'
 import KakaoMap from '@/src/shared/ui/KakaoMap'
 import type { CourseType } from '@/src/entities/course/type'
 import { formatDateToYYYYMMDD } from '@/src/shared/utils/date'
-import { CATEGORY } from '@/src/shared/entities/type'
+import { CATEGORY } from '@/src/entities/category/type'
 import { OptionHeader } from '@/src/widgets/header'
 
 const COURSE_PLAN = {
@@ -47,7 +47,7 @@ export default function CoursePlanLayout({
                 key={index}
                 className='px-[10px] py-[5px] text-[12px] text-white border rounded-[15px] bg-container-light-blue'
               >
-                {CATEGORY[category]}
+                {CATEGORY[category as keyof typeof CATEGORY]}
               </span>
             )
           })}
